renamed and added inductor correction
